+++ src/index.ts
@@ -1,13 +1,18 @@
+import { Client } from "@xmtp/xmtp-js";
 import Airstack from "./airstack";
+import { sendInvite } from "./xmtp";
 
-export { sendInvite } from "./xmtp";
-
-export default class OnChainFriendsSDK {
-  Airstack: Airstack;
+export default class OnChainFriendsSDK extends Airstack {
   /**
    * Instantiate the OnChainFriends SDK using the airstack api key
    */
   constructor(airstackApiKey: string) {
-    this.Airstack = new Airstack(airstackApiKey);
+    super(airstackApiKey);
+  }
+
+  async inviteUser(client: Client, toAddress: string, messageContent: string) {
+    return await sendInvite(client, toAddress, messageContent);
   }
 }
+
+export const FriendFetcher = OnChainFriendsSDK;