+++ src/playground.ts
@@ -1,13 +1,4 @@
-// import OnChainFriendsSDK from "./index.js";
+// import { Client } from "@xmtp/xmtp-js";
+// import OnChainFriendsSDK from ".";
 
 // const sdk = new OnChainFriendsSDK("118f7b49bcf804ce9af37bbfe3cad9f24");
-
-import Airstack from "./airstack/index";
-const sdk = new Airstack("118f7b49bcf804ce9af37bbfe3cad9f24");
-
-sdk
-  .getFarcasterFollowing("vitalik.eth")
-  .then((res) => {
-    console.dir(res, { depth: null });
-  })
-  .catch((err) => console.log(err));